aider: Added "New Post" button to `HomePage` component for logged-in users, linking to `NewPost` page.
# Aider chat conversation:

USER: I don't see the option to add the blog posts anywhere.

ASSISTANT: To provide the option to add blog posts, we need to implement a button or link that navigates to the page where users can create a new post. This should be visible only to logged-in users. I will add a "New Post" button to the `HomePage` component that links to the `NewPost` page.

Here's the diff to add the "New Post" button:

```diff
--- src/client/pages/Home.jsx
+++ src/client/pages/Home.jsx
@@ ... @@
 import { useQuery } from '@wasp/queries';
 import getPosts from '@wasp/queries/getPosts';
+import { useAuth } from '@wasp/auth';

 const HomePage = () => {
   const { data: posts, isLoading, error } = useQuery(getPosts);
+  const auth = useAuth();

   // Log the posts to the console for debugging
   React.useEffect(() => {
@@ ... @@
   return (
     <div className='p-4'>
       <div className='mb-4'>
+        {auth.user && (
+          <div className='mb-4'>
+            <Link to="/new-post" className='px-4 py-2 bg-blue-500 text-white rounded hover:bg-blue-600'>
+              New Post
+            </Link>
+          </div>
+        )}
         {posts.map((post) => (
           <div key={post.id} className='flex justify-between items-center p-2 bg-gray-100 mb-2 rounded-lg'>
             <Link to={`/post/${post.id}`} className='font-semibold'>{post.title}</Link>
             <span className='text-sm text-gray-600'>{post.author.username}</span>
           </div>
         ))}
       </div>
-      { /* Only show the New Post button if the user is logged in */ }
-      { /* TODO: Check if user is logged in and render button if true */ }
     </div>
   );
 };
```

This change checks if the user is logged in using the `useAuth` hook and conditionally renders the "New Post" button if a user is present. The button links to the `/new-post` route, which should be the page for creating a new post. Make sure that the `NewPost` page and the corresponding route are properly set up to handle post creation.