aider: Add links to the "New Post", "Login", and "Signup" pages in the header of the `Layout.jsx` file.
# Aider chat conversation:

USER: Add the existing pages to the layout

ASSISTANT: To add the existing pages to the layout, I will update the `Layout.jsx` file to include links to the "New Post", "Login", and "Signup" pages. Here is the diff for the changes:

```diff
--- src/client/Layout.jsx
+++ src/client/Layout.jsx
@@ ... @@
           <div>
             <Link to="/" className="text-xl2 font-semibold mr-4">Home</Link>
+            <Link to="/new-post" className="text-xl2 font-semibold mr-4">New Post</Link>
+            <Link to="/login" className="text-xl2 font-semibold mr-4">Login</Link>
+            <Link to="/signup" className="text-xl2 font-semibold">Signup</Link>
           </div>
           { user ? (
```

This diff adds links to the "New Post", "Login", and "Signup" pages in the header of the `Layout.jsx` file. Please note that the "New Post" link should only be visible to authenticated users, so you may want to wrap it with a conditional check for `user` similar to the logout button. If you need this conditional logic implemented, please let me know.